RM-55013 Refactor: introduce interface/implementation for EbicsService
